+++ .hlint.yaml
@@ -1,7 +1,7 @@
 ---
 - functions:
-  - {name: unsafePerformIO, within: [PlutusPrelude, Language.PlutusCore.TestSupport.Generator]}
-  - {name: error, within: [Main, Evaluation.CkMachine, Evaluation.Constant.Success, Language.PlutusCore.Constant.Apply, Language.PlutusCore.CkMachine, Language.PlutusCore.TestSupport.Generator, Language.PlutusCore.Constant.Make, Language.PlutusCore.TestSupport.TypedBuiltinGen, Language.PlutusCore.TH]}
+  - {name: unsafePerformIO, within: [PlutusPrelude, Language.PlutusCore.Generators.Internal.Entity]}
+  - {name: error, within: [Main, Evaluation.CkMachine, Evaluation.Constant.Success, Language.PlutusCore.Constant.Apply, Language.PlutusCore.CkMachine, Language.PlutusCore.Generators.Internal.Entity, Language.PlutusCore.Constant.Make, Language.PlutusCore.Generators.Internal.TypedBuiltinGen, Language.PlutusCore.TH]}
   - {name: undefined, within: [Language.PlutusCore.Constant.Apply]}
   - {name: fromJust, within: []}
   - {name: foldl, within: []}
@@ -18,7 +18,7 @@
 - error: {lhs: "mconcat", rhs: "fold", name: "Generalize mconcat"}
 
 - ignore: {name: Reduce duplication, within: [Language.PlutusCore.Renamer, Language.PlutusCore.Constant.Prelude, Language.PlutusCore.StdLib.Data.Bool, Language.PlutusCore.StdLib.Data.ChurchNat, Language.PlutusCore.StdLib.Data.Function, Language.PlutusCore.StdLib.Data.List, Language.PlutusCore.StdLib.Data.Nat]}
-- ignore: {name: Redundant $, within: [Evaluation.Constant.Success, Language.PlutusCore.TestSupport.TypedBuiltinGen]}
+- ignore: {name: Redundant $, within: [Evaluation.Constant.Success, Language.PlutusCore.Generators.Internal.TypedBuiltinGen]}
 # this is rarely an improvement, also ignored in cardano
 - ignore: {name: Move brackets to avoid $}
