+++ language-plutus-core/language-plutus-core.cabal
@@ -38,7 +38,6 @@ library
         Language.PlutusCore.Quote
         Language.PlutusCore.CkMachine
         Language.PlutusCore.Constant
-        Language.PlutusCore.TestSupport
         Language.PlutusCore.View
         Language.PlutusCore.StdLib.Data.Bool
         Language.PlutusCore.StdLib.Data.ChurchNat
@@ -47,9 +46,10 @@ library
         Language.PlutusCore.StdLib.Data.Nat
         Language.PlutusCore.StdLib.Data.Unit
         Language.PlutusCore.StdLib.Type
+        Language.PlutusCore.Generators
         PlutusPrelude
     build-tools: alex -any, happy >=1.17.1
-    hs-source-dirs: src prelude stdlib src
+    hs-source-dirs: src prelude stdlib generators
     other-modules:
         Language.PlutusCore.Type
         Language.PlutusCore.Name
@@ -63,14 +63,14 @@ library
         Language.PlutusCore.Error
         Language.PlutusCore.TypeSynthesis
         Language.PlutusCore.Normalize
-        Language.PlutusCore.TestSupport.Denotation
-        Language.PlutusCore.TestSupport.Generator
-        Language.PlutusCore.TestSupport.TypedBuiltinGen
-        Language.PlutusCore.TestSupport.Utils
         Language.PlutusCore.CBOR
         Language.PlutusCore.PrettyCfg
         Language.PlutusCore.Subst
         Language.PlutusCore.TH
+        Language.PlutusCore.Generators.Internal.Denotation
+        Language.PlutusCore.Generators.Internal.Entity
+        Language.PlutusCore.Generators.Internal.TypedBuiltinGen
+        Language.PlutusCore.Generators.Internal.Utils
         Data.Functor.Foldable.Monadic
         Data.Text.Prettyprint.Doc.Custom
     default-language: Haskell2010
